move items out of directories that dont need to be
