+++ src/App.vue
@@ -6,7 +6,7 @@
 </template>
 
 <script>
-import Navigation from '@/components/Navigation/Navigation.vue'
+import Navigation from '@/components/Navigation.vue'
 
 export default {
   name: 'App',
