+++ src/views/Stats.vue
@@ -25,7 +25,7 @@ export default {
   },
   methods: {
     showStats: function () {
-      fetch(`/php/Stats/stats.php`)
+      fetch(`/php/stats.php`)
       .then(response => {
         if (response.status !== 200) {
           return []
